fix(email): Improvements and django-ses updated
