+++ cl/users/admin.py
@@ -6,7 +6,13 @@ from cl.alerts.admin import AlertInline, DocketAlertInline
 from cl.donate.admin import DonationInline, MonthlyDonationInline
 from cl.favorites.admin import FavoriteInline, UserTagInline
 from cl.lib.admin import AdminTweaksMixin
-from cl.users.models import BackoffEvent, BarMembership, EmailFlag, UserProfile
+from cl.users.models import (
+    BackoffEvent,
+    BarMembership,
+    EmailFlag,
+    EmailSent,
+    UserProfile,
+)
 
 
 def get_email_confirmed(obj):
@@ -71,6 +77,16 @@ class BackoffEventAdmin(admin.ModelAdmin):
         "email_address",
         "id",
         "retry_counter",
+        "date_created",
+    )
+
+
+@admin.register(EmailSent)
+class EmailSentAdmin(admin.ModelAdmin):
+    list_display = (
+        "to",
+        "id",
+        "subject",
         "date_created",
     )
